requirements.txt
fastapi
uvicorn
torch
transformers
scikit-learn
numpy==1.24.0
pymongo
openai
python-dotenv
pydantic